fix types in isPrime

+++ spcppl/numbers/primes.hpp
@@ -14,7 +14,7 @@ bool isPrime(int n) {
 }
 
 bool isPrime(int64_t n) {
-	for (int i = 2; i * i <= n; ++i) {
+	for (int64_t i = 2; i * i <= n; ++i) {
 		if (n % i == 0) {
 			return false;
 		}
